pretrain_data/s2/v2/requirements.txt
pandas
smashed[remote]>=0.19.3
torch==1.13.1
springs
s3fs
fsspec
pycld3
cached_path
blingfire
pyarrow
